// Frontend/src/app/model/x_auftrag_ext.ts
export class XAuftragExt {
    name: string;
    EO_ID: string;
    ORDER_ID_CRM: string;
    ORDER_ID_BSSOE: string;
    ORDER_ID_SMF: string;
    ORDER_ID_DKK: string;
    ORDER_ID_SNR: string;
    ORDER_ID_MISC: string;
    TS_CREATED: string;
    TS_LAST_UPDATE: string;
    URSPRUNGS_KWT: string;
    LAST_UPD: string;
    constructor() {
        console.log("class XAuftragExt");
        this.name='XAuftragExt';
        this.EO_ID='';
        this.ORDER_ID_CRM='';
        this.ORDER_ID_BSSOE='';
        this.ORDER_ID_SMF='';
        this.ORDER_ID_DKK='';
        this.ORDER_ID_SNR='';
        this.ORDER_ID_MISC='';
        this.TS_CREATED='';
        this.TS_LAST_UPDATE='';
        this.URSPRUNGS_KWT='';
        this.LAST_UPD='';
        }
}